// server-backup/front-end/src/service/api/student/type.ts
export type StudentEnrollDTO = {
  student_id: number;
  course_section_id: number;
  document_language: "th" | "en";

  company_register_number: string;
  company_name_en: string;
  company_name_th: string;
  company_address: string;
  company_map: string;
  company_email: string;
  company_phone_number: string;
  company_type: string;

  picture_1: string | File;
  picture_2: string | File;

  start_date: Date | string;
  end_date: Date | string;
  coordinator: string;
  coordinator_phone_number: string;
  coordinator_email: string;
  supervisor: string;
  supervisor_phone_number: string;
  supervisor_email: string;
  department: string;
  position: string;
  job_description: string;
};